+++ saleor/graphql/checkout/types.py
@@ -611,11 +611,6 @@ class Checkout(SyncWebhookControlContextModelObjectType[models.Checkout]):
     customer_note = graphene.String(
         required=True, description=f"The customer note for the checkout. {ADDED_IN_321}"
     )
-    note = graphene.String(
-        required=True,
-        description="The note for the checkout.",
-        deprecation_reason="Use `customerNote` instead.",
-    )
     discount = graphene.Field(
         Money,
         description=(