+++ saleor/graphql/product/types/products.py
@@ -95,9 +95,6 @@ from ...core.types import (
 )
 from ...core.types.context import ChannelContextType
 from ...core.utils import from_global_id_or_error
-from ...core.validators import (
-    validate_one_of_args_is_in_query,
-)
 from ...meta.types import ObjectWithMetadata
 from ...order.dataloaders import (
     OrderByIdLoader,
@@ -1037,13 +1034,6 @@ class Product(ChannelContextType[models.Product]):
         description="Get a single product image by ID.",
         deprecation_reason="Use the `mediaById` field instead.",
     )
-    variant = graphene.Field(
-        ProductVariant,
-        id=graphene.Argument(graphene.ID, description="ID of the variant."),
-        sku=graphene.Argument(graphene.String, description="SKU of the variant."),
-        description="Get a single variant by SKU or ID.",
-        deprecation_reason="Use top-level `variant` query.",
-    )
     variants = NonNullList(
         ProductVariant,
         description=(
@@ -1482,40 +1472,6 @@ class Product(ChannelContextType[models.Product]):
     @staticmethod
     def resolve_images(root: ChannelContext[models.Product], info):
         return ImagesByProductIdLoader(info.context).load(root.node.id)
-
-    @staticmethod
-    def resolve_variant(root: ChannelContext[models.Product], info, id=None, sku=None):
-        validate_one_of_args_is_in_query("id", id, "sku", sku)
-
-        def get_product_variant(
-            product_variants,
-        ) -> ProductVariant | None:
-            if id:
-                id_type, variant_id = graphene.Node.from_global_id(id)
-                if id_type != "ProductVariant":
-                    return None
-
-                return next(
-                    (
-                        variant
-                        for variant in product_variants
-                        if variant.node.id == int(variant_id)
-                    ),
-                    None,
-                )
-            if sku:
-                return next(
-                    (
-                        variant
-                        for variant in product_variants
-                        if variant.node.sku == sku
-                    ),
-                    None,
-                )
-            return None
-
-        variants = Product.resolve_variants(root, info)
-        return variants.then(get_product_variant)
 
     @staticmethod
     def resolve_variants(root: ChannelContext[models.Product], info):